Initial infrastructure for RFB integration
This is the first of a series of patches to enable RFB support in
Velocity. Currently, this enables the program 'VNC Viewer' to connect
and attach to the latest virtual machine.

ScreenCaptureKit has been removed due to it not being functional when
the window is hidden away because MacOS will decrease its refresh rate.